+++ src/Rules/PaperExistsRule.php
@@ -7,6 +7,7 @@ namespace JacobJoergensen\LaravelPaper\Rules;
 use Closure;
 use Illuminate\Contracts\Validation\ValidationRule;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Translation\PotentiallyTranslatedString;
 
 final readonly class PaperExistsRule implements ValidationRule
 {
@@ -18,13 +19,16 @@ final readonly class PaperExistsRule implements ValidationRule
         private string $column,
     ) {}
 
+    /**
+     * @param  Closure(string, ?string=): PotentiallyTranslatedString  $fail
+     */
     public function validate(string $attribute, mixed $value, Closure $fail): void
     {
         $query = $this->model::query();
         $exists = $query->where($this->column, $value)->exists();
 
         if (! $exists) {
-            $fail("The selected $attribute does not exist.");
+            $fail('validation.exists')->translate();
         }
     }
 }